feat(flow): add async request handling
